requirements.txt
--find-links https://download.pytorch.org/whl/cu116
#torch==1.13.1+cu116
torch==1.13.1

pytorch-lightning==1.9.4
transformers==4.26.1
lightning-transformers==0.2.5

hydra-core==1.3.2
wandb==0.14.0
scipy==1.10.1
numba==0.56.4
tqdm